I think the problem now is with checksum.coffee
The file is getting successfully downloaded, verified by
commenting out the clearing of the filesystem.

It seems that when trying to read the stream in checksum.coffee
there's not data, I guess possibly because it is trying to stream
a folder?

I've gotten as far as I could right now, I'm doing this with a
sixpack of beer and I think I'm over the Balmer Peak.

Sorry for the mess, making a PR instead of just commiting...